Checkup: Move the code creating the VMIs to a new file
Currently, the code that creates the two custom VMIs
is located in `checkup.go`.

Add a new `vmi.go` file in the `checkup` package, that will contain all
the logic related to creating customized VMIs.
